Cleanup test suite
