+++ tests/ApplicationTest.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace WordPlate\Tests;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Dotenv\Dotenv;
 use WordPlate\Application;
 
 class ApplicationTest extends TestCase
@@ -24,11 +25,11 @@ class ApplicationTest extends TestCase
      */
     public function testBasePath()
     {
-        $application = new Application(__DIR__ . '/stubs');
+        $application = new Application(__DIR__);
         $application->run();
 
         $this->assertInstanceOf(Application::class, $application);
-        $this->assertSame(__DIR__ . '/stubs', $application->getBasePath());
+        $this->assertSame(__DIR__, $application->getBasePath());
     }
 
     /**
@@ -37,12 +38,38 @@ class ApplicationTest extends TestCase
      */
     public function testPublicPath()
     {
-        $application = new Application(__DIR__ . '/stubs');
+        $application = new Application(__DIR__);
         $application->run();
 
-        $this->assertSame(__DIR__ . '/stubs/public', $application->getPublicPath());
+        $this->assertSame(__DIR__ . '/public', $application->getPublicPath());
 
         $application->setPublicPath(__DIR__);
         $this->assertSame(__DIR__, $application->getPublicPath());
+    }
+
+    public function testEnvFunction()
+    {
+        $this->assertSame('mcfly', env('DOTENV_DEFAULT', 'mcfly'));
+
+        $variables = [
+            ['DOTENV_FALSE', 'false', false],
+            ['DOTENV_FALSE', '(false)', false],
+            ['DOTENV_TRUE', 'true', true],
+            ['DOTENV_TRUE', '(true)', true],
+            ['DOTENV_EMPTY', 'empty', ''],
+            ['DOTENV_EMPTY', '(empty)', ''],
+            ['DOTENV_NULL', 'null', null],
+            ['DOTENV_NULL', '(null)', null],
+            ['DOTENV_QUOTED', '"null"', 'null'],
+            ['DOTENV_QUOTED', "'null'", 'null'],
+            ['DOTENV_STRING', 'einstein', 'einstein'],
+        ];
+
+        $environment = new Dotenv();
+
+        foreach ($variables as [$key, $value, $expected]) {
+            $environment->populate([$key => $value]);
+            $this->assertSame($expected, env($key));
+        }
     }
 }
