--- tests/stubs/public/wordpress/wp-includes/plugin.php
@@ -1,27 +1,0 @@
-<?php
-
-/**
- * Copyright (c) Vincent Klaiber.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- *
- * @see https://github.com/wordplate/framework
- */
-
-declare(strict_types=1);
-
-function add_action($name, $callback)
-{
-    $callback();
-}
-
-function add_filter()
-{
-    //
-}
-
-function remove_filter()
-{
-    //
-}
